Disable stable function evaluation in pl/pgsql simple expressions
In Greenplum, stable functions are evaluated during planning to
facilitate partition elimination.  Simple expressions in pl/pgsql UDFs
cannot eliminate partitions because they do not involve table scans.
This patch disables stable function evaluation in the planner in the
case that we are planning what pl/pgsql has previously determined to be
a simple expression.

Without this patch, every time a simple expression is evaluated, a new
plan and a new ExprState are generated.  Because we only clean up the
expression states created for simple expressions at the end of
transaction, this will result in millions of expression states, one per
tuple, being generated that are not reused after they are initially
created.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>
Co-authored-by: Asim R P <apraveen@pivotal.io>
Co-authored-by: Joao Pereira <jdealmeidapereira@pivotal.io>