Add the initial tests for lwlock
Added test for:
1. NumLWLocks
2. LWLockShmemSize
3. LWLockAcquire
4. LWLockRelease
6. LWLockHeldByMe
7. LWLockHeldExclusiveByMe

The first 2 tests are just to make sure we have enough coverage when we
create the fake of lwlock. It will be helpful to ensure that the values
returned from these functions is correct as well

In this patch a new `expected_elog_with_message` function was added to
the elog mocking helper to do the setup of the mock that ensure that
elog is called with the correct format.

This is where the Shared Memory fake starts to pay off. In this commit
we add more functionality to our Fake to allow us to retrieve some
"inner" information about the module itself. The functions
`shmem_fake_get_total_memory_allocated` and
`shmem_fake_retrieve_allocated_memory` allow us to look into the
amount of memory allocated by lwlock and also to return a pointer to
that shared memory.

We are using this trick to allow us to manipulate the lwlock module
information. This information is used to be able to test that a lock
can only be aquired if no other process as it. There is more
information about this in the comments of the functions that allow us
to have this behavior.